+++ src/Symfony/Component/Tui/CHANGELOG.md
@@ -5,6 +5,8 @@ CHANGELOG
 ---
 
  * Add `AbstractWidget::attachChild()` and `AbstractWidget::detachChild()` to wire child widgets
+ * Add multi-select support to `SelectListWidget`
+ * [BC BREAK] Add `$multiselect` as the third argument of `SelectListWidget::__construct()`, moving `$keybindings` to fourth position
 
 8.1
 ---